Migrate logic from transferinject to denommetadata
