+++ x/denommetadata/types/expected_keepers.go
@@ -5,6 +5,8 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
 	transfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
+
+	rollapptypes "github.com/dymensionxyz/dymension/v3/x/rollapp/types"
 )
 
 // TransferKeeper defines the expected transfer keeper
@@ -15,10 +17,19 @@ type TransferKeeper interface {
 
 // BankKeeper defines the expected interface needed
 type BankKeeper interface {
-	HasDenomMetaData(ctx sdk.Context, denom string) bool
+	GetDenomMetaData(ctx sdk.Context, denom string) (types.Metadata, bool)
 	SetDenomMetaData(ctx sdk.Context, denomMetaData types.Metadata)
 }
 
 type DenomMetadataKeeper interface {
 	CreateDenomMetadata(ctx sdk.Context, metadata types.Metadata) error
+}
+
+type RollappKeeper interface {
+	SetRollapp(ctx sdk.Context, rollapp rollapptypes.Rollapp)
+	ExtractRollappFromChannel(
+		ctx sdk.Context,
+		rollappPortOnHub string,
+		rollappChannelOnHub string,
+	) (*rollapptypes.Rollapp, error)
 }
