--- x/transferinject/types/codec.go
@@ -1,8 +1,0 @@
-package types
-
-import (
-	"github.com/cosmos/cosmos-sdk/codec"
-	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
-)
-
-var ModuleCdc = codec.NewProtoCodec(cdctypes.NewInterfaceRegistry())
